Update ts0601_din_power.py, fix invalid type.
Prevent "Summation received" being stuck at 655.36kWh. Side bonus: no more log padding with information about 65537 not being a valid uint16.

+++ zhaquirks/tuya/ts0601_din_power.py
@@ -173,7 +173,7 @@ class HikingManufClusterDinPower(TuyaManufClusterAttributes):
         )
         energy_received: Final = ZCLAttributeDef(
             id=HIKING_TOTAL_ENERGY_RECEIVED_ATTR,
-            type=t.uint16_t,
+            type=t.uint32_t,
             is_manufacturer_specific=True,
         )
         voltage_current: Final = ZCLAttributeDef(
